+++ NEWS.md
@@ -74,7 +74,14 @@
 
  * New example `examples/target-runner-python/target-runner-python-win.bat`
    contributed by Levi Ribeiro.
-   
+
+ * `path_rel2abs()` also searches in system paths.     (Manuel López-Ibáñez)
+
+ * New scenario options `targetRunnerLauncher` and `targetRunnerLauncherArgs`
+   to help in cases where the target-runner must be invoked via another
+   software with particular options (such as `python.exe` in Windows).
+                                                       (Manuel López-Ibáñez)
+
 
 # irace 3.4.1  (31/03/2020)